Create Cash
In cash.c, I've implemented a program that prompts the user for the number of cents that a customer is owed and then prints the smallest number of coins with which that change can be made. It's several functions.
One of those functions, get_cents, takes no arguments (as indicated by void) and returns an int. The rest of the functions all take one argument, an int, and also return an int. . 

get_cents Implemented in such a way that the function prompts the user for a number of cents using get_int (from CS50's library) and then returns that number as an int. If the user inputs a negative int, the code prompts the user again.

calculate_quarters Implemented in such a way that the function calculates (and returns as an int) how many quarters a customer should be given if they’re owed some number of cents. For instance, if cents is 25, then calculate_quarters should return 1.

If cents is 26 or 49 (or anything in between, then calculate_quarters also returns 1. 

If cents is 50 or 74 (or anything in between), then calculate_quarters returns 2. And so forth.

calculate_dimes Implemented in such a way that the function calculates the same for dimes.
calculate_nickels Implemented in such a way that the function calculates the same for nickels.
calculate_pennies Implemented in such a way that the function calculates the same for pennies.